需求：用django rest framework 的APIView和ModelSerializer 更新内容： refactor(重構): 1、修改/apps/music/views.py 文件路径
